+++ Assets/Monsters/TargetLocator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 622a14701c7210745bc60c40451afb57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
